fix: shift add_custom_field_for_customer_suppliers_groups to custom_fields folder and automate tax_category values
