+++ csf_tz/patches.txt
@@ -13,7 +13,7 @@ csf_tz.patches.custom_fields.payroll_approval_custom_fields
 csf_tz.patches.custom_fields.attendance_overtime_calculation_custom_fields
 csf_tz.patches.add_custom_fields_for_employee_advance
 csf_tz.patches.add_custom_for_employee_doctype
-csf_tz.patches.add_custom_field_for_cusomer_suppliers_groups
+csf_tz.patches.custom_fields.add_custom_field_for_customer_suppliers_groups
 csf_tz.patches.fix_module_for_core_reports
 csf_tz.patches.create_the_stock_entry_type
 csf_tz.patches.custom_fields.add_fields_in_stock_entry
